DRAAdminAccess: add example (#112)
* DRAAdminAccess: add example


* address comments and make it os agnostic


* address comments


* address comments


---------

+++ test/e2e/e2e.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-# Copyright 2024 The Kubernetes Authors.
+# Copyright The Kubernetes Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -52,6 +52,8 @@ timeout --foreground 15s bash -c verify-webhook
 kubectl create -f demo/gpu-test1.yaml
 kubectl create -f demo/gpu-test2.yaml
 kubectl create -f demo/gpu-test3.yaml
+# deploying this earlier to ensure the pod can access in-use devices and does not block future allocations of the same devices
+kubectl create -f demo/gpu-test7.yaml 
 kubectl create -f demo/gpu-test4.yaml
 kubectl create -f demo/gpu-test5.yaml
 kubectl create -f demo/gpu-test6.yaml
@@ -432,6 +434,23 @@ if [[ "$gpu_test6_pod0_ctr0_timeslice_interval" != "Default" ]]; then
   exit 1
 fi
 
+
+kubectl wait --for=condition=Ready -n gpu-test7 pod/pod0 --timeout=120s
+gpu_test_7=$(kubectl get pods -n gpu-test7 | grep -c 'Running')
+if [ $gpu_test_7 != 1 ]; then
+    echo "gpu_test_7 $gpu_test_7 failed to match against 1 expected pod"
+    exit 1
+fi
+
+gpu_test7_pod0_ctr0_logs=$(kubectl logs -n gpu-test7 pod0 -c ctr0)
+gpu_test7_pod0_ctr0_admin_access=$(echo "$gpu_test7_pod0_ctr0_logs" | sed -nE "s/^declare -x DRA_ADMIN_ACCESS=\"(.+)\"$/\1/p")
+if [[ "$gpu_test7_pod0_ctr0_admin_access" != "true" ]]; then
+  echo "Expected Pod gpu-test7/pod0, container ctr0 to have DRA_ADMIN_ACCESS=true, got $gpu_test7_pod0_ctr0_admin_access"
+  exit 1
+fi
+echo "Pod gpu-test7/pod0, container ctr0 has admin access: $gpu_test7_pod0_ctr0_admin_access"
+
+
 # test that deletion is fast (less than the default grace period of 30s)
 # see https://github.com/kubernetes/kubernetes/issues/127188 for details
 kubectl delete -f demo/gpu-test1.yaml --timeout=25s
@@ -440,8 +459,9 @@ kubectl delete -f demo/gpu-test3.yaml --timeout=25s
 kubectl delete -f demo/gpu-test4.yaml --timeout=25s
 kubectl delete -f demo/gpu-test5.yaml --timeout=25s
 kubectl delete -f demo/gpu-test6.yaml --timeout=25s
+kubectl delete -f demo/gpu-test7.yaml --timeout=25s
 
-# # Webhook should reject invalid v1 resources
+# Webhook should reject invalid v1 resources
 if ! kubectl create --dry-run=server -f- <<'EOF' 2>&1 | grep -qF 'unknown time-slice interval'
 apiVersion: resource.k8s.io/v1
 kind: ResourceClaim
@@ -473,7 +493,7 @@ then
   exit 1
 fi
 
-# # Webhook should reject invalid v1beta1 resources
+# Webhook should reject invalid v1beta1 resources
 if ! kubectl create --dry-run=server -f- <<'EOF' 2>&1 | grep -qF 'unknown time-slice interval'
 apiVersion: resource.k8s.io/v1beta1
 kind: ResourceClaim
@@ -501,6 +521,7 @@ then
   exit 1
 fi
 
+# Webhook should reject invalid v1 ResourceClaimTemplates
 if ! kubectl create --dry-run=server -f- <<'EOF' 2>&1 | grep -qF 'unknown time-slice interval'
 apiVersion: resource.k8s.io/v1
 kind: ResourceClaimTemplate
